+++ include/timer-proc.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#include "state.hpp"
-
-#include <rain.hpp>
-
-namespace KbAsMouse {
-	VOID CALLBACK TimerProc(HWND hWnd, UINT uMsg, UINT_PTR idEvent, DWORD time);
-}